temp relation

+++ src/idea/idea.service.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { throws } from 'assert';
 import { UserEntity } from 'src/user/user.entity';
 import { Repository } from 'typeorm';
-import { IdeaDTO } from './idea.dto';
+import { IdeaDTO, IdeaRO } from './idea.dto';
 import { IdeaEntity } from './idea.entity';
 
 @Injectable()
@@ -13,47 +13,62 @@ export class IdeaService {
         @InjectRepository(UserEntity) private userRepository: Repository<UserEntity>
     ) {}
 
-    async showAll() {
-        return await this.ideaRepository.find({ relations: ['author']});
+    async showAll(): Promise<IdeaRO[]> {
+        const ideas =  await this.ideaRepository.find({ relations: ['author'] });
+        return ideas.map(idea => this.toResponseObject(idea));
     }
 
-    async create(userId: string, data: IdeaDTO) {
+    async create(userId: string, data: IdeaDTO): Promise<IdeaRO> {
         console.log('userId', userId);
         
         // const user = await this.userRepository.findOne({ where: { id: userId } });
-        const user = await this.userRepository.findOne({ where: { id: 4 } });
+        const user = await this.userRepository.findOne({ where: { id: 1 } });
        
         const idea = await this.ideaRepository.create({...data, author: user});
         await this.ideaRepository.save(idea);
-        return {...idea, author: idea.author.toResponseObject() };
+        return this.toResponseObject(idea);
     }
 
-    async read(id: string) {
-        const idea = await this.ideaRepository.findOne({ where: { id }})
+    async read(id: string): Promise<IdeaRO> {
+        const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] })
         if(!idea) {
             throw new HttpException('Not found idea', HttpStatus.NOT_FOUND);
         }
 
-        return idea;
+        return this.toResponseObject(idea);
     }
 
-    async update(id: string, data: Partial<IdeaDTO>) {
-        const idea = await this.ideaRepository.findOne({ where: { id }});
+    async update(id: string, userId: string, data: Partial<IdeaDTO>): Promise<IdeaRO> {
+        const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
         if(!idea) {
             throw new HttpException('Not found idea', HttpStatus.NOT_FOUND);
         }
 
+        this.ensureOwnerShip(idea, userId);
+        
         await this.ideaRepository.update({id}, data);
-        return idea;
+        return this.toResponseObject(idea);
     }
 
-    async destroy(id: string) {
-        const idea = await this.ideaRepository.findOne({ where: { id }});
+    async destroy(id: string, userId: string) {
+        const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
         if(!idea) {
             throw new HttpException('Not found idea', HttpStatus.NOT_FOUND);
         }
+
+        this.ensureOwnerShip(idea, userId);
 
         await this.ideaRepository.delete({ id });
         return idea;
+    }
+
+    private toResponseObject(idea: IdeaEntity): IdeaRO {
+        return {...idea, author: idea.author.toResponseObject(false) };
+    }
+
+    private ensureOwnerShip(idea: IdeaEntity, userId: string) {
+        if(idea.author.id !== userId) {
+            throw new HttpException('Incorrect user', HttpStatus.UNAUTHORIZED);
+        }
     }
 }
